test(cli): move the soucemap test into `serializer-plugin.test.ts`
