+++ packages/@expo/cli/e2e/__tests__/export/serializer-plugins.test.ts
@@ -1,6 +1,8 @@
 /* eslint-env jest */
+import JsonFile from '@expo/json-file';
 import execa from 'execa';
 import fs from 'fs';
+import type { BasicSourceMap } from 'metro-source-map';
 import path from 'path';
 
 import { runExportSideEffects } from './export-side-effects';
@@ -57,5 +59,21 @@ describe('exports with serializer plugins', () => {
 
       expect(fs.existsSync(path.join(outputDir, mapFile!))).toBe(true);
     }
+  });
+
+  it('source maps do not contain project paths', async () => {
+    const files = findProjectFiles(outputDir).filter((file) => file.endsWith('.map'));
+
+    // Ensure the source map is available
+    expect(files).not.toEqual([]);
+
+    // Load the sourcemap and parse it as JSON
+    const rawmap = await JsonFile.readAsync(path.resolve(outputDir, files[0]));
+    const sourcemap = rawmap as unknown as BasicSourceMap;
+
+    // Ensure the sourcemap does not contain absolute paths to the project directory
+    expect(sourcemap.sources).not.toEqual(
+      expect.arrayContaining([expect.stringContaining(projectRoot)])
+    );
   });
 });
